Data.CAS.RemoteCache

Description

This module defines the remote caching mechanism of funflow which is used to keep several funflow stores (possibly on different machines) in sync.

Documentation

class Monad m => Cacher m a where

A simple mechanism for remote-caching.

Provides a way to push a path to the cache and pull it back.

No assumption is made on the availability of a store path. In particular, pushing a path to the cache doesn't mean that we can pull it back.

Methods

push

Arguments

:: a

-> ContentHash
Primary key: hash of the content

-> Maybe ContentHash
Secondary key: hash of the dependencies

-> Path Abs Dir
Path to the content

-> m PushResult

pull :: a -> ContentHash -> Path Abs Dir -> m (PullResult ())

data PullResult a

The result of a tentative pull from the remote cache

Constructors

PullOK a

NotInCache

PullError String

data PushResult

The result of a tentative push to the remote cache

Constructors

PushOK

PushError String

data AliasResult

Constructors

AliasOK

TargetNotInCache

AliasError String

data NoCache

A dummy remote cache implementation which does nothing

Constructors

NoCache

memoryCache :: MonadIO m => m MemoryCache

pullAsArchive :: MonadIO m => (ContentHash -> m (PullResult ByteString)) -> ContentHash -> Path Abs Dir -> m (PullResult ())

pushAsArchive

Arguments

:: MonadIO m

=> (ContentHash -> ContentHash -> m (Either String ()))
How to create the aliases

-> (ContentHash -> ByteString -> m PushResult)
How to push the content

-> ContentHash
Primary key

-> Maybe ContentHash
Secondary key

-> Path Abs Dir

-> m PushResult

Push the path as an archive to the remote cache